+++ src/reduce_fsm/fsm.clj
@@ -6,9 +6,21 @@ This package allows you to:
  - Create lazy sequences from state machines (see fsm-seq)
  - Create stateful filter functions for use with filter/remove (see fsm-filter)"
   (:require
-    [clojure.core.match :refer [match]]
+    [cljs.core.match]
     [clojure.core.match.regex]
-    [clojure.set :as set]))
+    [clojure.core.match]
+    [clojure.set :as set]
+    [reduce-fsm.fsm.impl :as impl]))
+
+;; I'm grateful to [1].
+;;
+;; [1]: https://code.thheller.com/blog/shadow-cljs/2019/10/12/clojurescript-macros.html
+;;
+(defmacro match
+  [vars & clauses]
+  (if (:ns &env)
+    `(cljs.core.match/match ~vars ~@clauses)
+    `(clojure.core.match/match ~vars ~@clauses)))
 
 (defn- fsm-fn?
   "return true if the symbol will be treated as a function in fsm state definitions."
@@ -92,23 +104,11 @@ This package allows you to:
     (sanity-check-fsm state-maps)
     state-maps))
 
-(defn- state-fn-name
-  "Create a name for the internal function that will represent this state.
-   We want it to be recognisable to the user so stack traces are more intelligible"
-  [sym]
-  (cond
-    (fn? sym) (let [fn-name (-> sym meta :name str)]
-                (if (empty? fn-name)
-                  (str (gensym "fn-"))
-                  fn-name))
-    (keyword? sym) (name sym)
-    :else (str sym)))
-
 (defn- state-fn-symbol
   "Create a name for the internal function that will represent this state.
    We want it to be recognisable to the user so stack traces are more intelligible"
   [sym]
-  (gensym (str "state" "-" (state-fn-name sym) "-")))
+  (gensym (str "state" "-" (impl/state-fn-name sym) "-")))
 
 (defn- state-for-action
   [state]
@@ -191,12 +191,6 @@ Parameters:
   [fsm-type state-maps]
   {::fsm-type fsm-type
    ::states (vec (map state-metadata state-maps))})
-
-(defn lookup-state
-  [state-fn-map the-state]
-  (if-let [a-state-fn (get state-fn-map the-state)]
-    a-state-fn
-    (throw (RuntimeException. (str "Could not find the state \"" the-state "\"")))))
 
 ;; ===================================================================================================
 ;; We want to turn an fsm definition looking like this:
@@ -286,7 +280,7 @@ See https://github.com/cdorrat/reduce-fsm for examples and documentation"
            ([acc# events#]
             (the-fsm# ~(-> state-maps first :from-state) acc# events#))
            ([initial-state# acc# events#]
-            (trampoline (lookup-state ~state-fn-map initial-state#) acc# events#)))
+            (trampoline (impl/lookup-state ~state-fn-map initial-state#) acc# events#)))
          ~(fsm-metadata :fsm state-maps)))))
 
 (defmacro defsm
@@ -297,9 +291,6 @@ See https://github.com/cdorrat/reduce-fsm for examples and documentation"
 
 ;; ===================================================================================================
 ;; support for incremental fsms
-(defn- state-disp-name
-  [sym]
-  (keyword (state-fn-name sym)))
 
 (defn- expand-inc-evt-dispatch
   "Expand the dispatch of a single event for incremental fsms, this corresponds to a single case in the match expression.
@@ -317,7 +308,7 @@ Parameters:
       (let [~new-acc ~(if (:action evt-map)
                         `(~(:action evt-map) ~acc ~evt ~(state-for-action from-state) ~(state-for-action (:to-state evt-map)))
                         acc)]
-        {:state ~(state-disp-name (:to-state evt-map))
+        {:state ~(impl/state-disp-name (:to-state evt-map))
          :value ~new-acc
          :fsm ~(state-fn-map (:to-state evt-map))
          :is-terminated? ~(if (fsm-fn? (:to-state evt-map)) ; if the target state is a fn exit on truthy value
@@ -336,7 +327,7 @@ Parameters:
     `(~this-state-fn  [~acc ~evt]
                       (~@(expand-dispatch dispatch-type evt acc)
                        ~@(mapcat (partial expand-inc-evt-dispatch state-fn-map state-params (:from-state state)  evt acc events) (:transitions state))
-                       :else  {:state ~(state-disp-name (:from-state state))
+                       :else  {:state ~(impl/state-disp-name (:from-state state))
                                :is-terminated? false
                                :value ~acc
                                :fsm ~this-state-fn}))))
@@ -362,10 +353,10 @@ Subsequent chained calls to  fsm-event will move the fsm thought it's states.
            ([] (the-fsm# ~default-acc))
            ([acc#] (the-fsm# ~(-> state-maps first :from-state) acc#))
            ([initial-state# acc#]
-            {:state (~state-disp-name initial-state#)
+            {:state (impl/state-disp-name initial-state#)
              :is-terminated? false
              :value acc#
-             :fsm (lookup-state ~state-fn-map initial-state#)}))
+             :fsm (impl/lookup-state ~state-fn-map initial-state#)}))
          ~(fsm-metadata :inc-fsm state-maps)))))
 
 (defn fsm-event
@@ -502,7 +493,7 @@ Example:
            ([acc#]
             (filter-builder# ~(-> state-maps first :from-state) acc#))
            ([initial-state# acc#]
-            (let [curr-state# (atom ((lookup-state ~state-fn-map initial-state#) acc#))]
+            (let [curr-state# (atom ((impl/lookup-state ~state-fn-map initial-state#) acc#))]
               (fn [evt#]
                 (let [[pass# next-state#] (@curr-state# evt#)]
                   (reset! curr-state# next-state#)
@@ -517,27 +508,6 @@ Example:
 
 ;; ===================================================================================================
 ;; fsm-seq impl
-
-(defn- next-emitted
-  "Process events with the fsm-seq function f until we emit a new value for the sequence"
-  [f]
-  (when f
-    (loop [[emitted next-step] (f)]
-      (if next-step
-        (if (not= ::no-event emitted)
-          [emitted next-step]
-          (recur (next-step)))
-        [emitted nil]))))
-
-(defn ^{:skip-wiki true} fsm-seq-impl*
-  "Create a lazy sequence from a fsm-seq state function"
-  [f]
-  (let [[emitted next-step] (next-emitted f)]
-    (lazy-seq
-      (if next-step
-        (cons emitted (fsm-seq-impl* next-step))
-        (when (not= ::no-event emitted)
-          (cons emitted nil))))))
 
 (defn- expand-seq-evt-dispatch
   "Expand the dispatch line for a single fsm-seq dispatch line.
@@ -670,7 +640,7 @@ See https://github.com/cdorrat/reduce-fsm for examples and documentation"
             (fsm-seq-fn# ~(-> state-maps first :from-state) acc# events#))
            ([initial-state# acc# events#]
             (when (seq events#)
-              (fsm-seq-impl* ((lookup-state ~state-fn-map initial-state#)  acc# events#)))))
+              (impl/fsm-seq-impl* ((impl/lookup-state ~state-fn-map initial-state#)  acc# events#)))))
          ~(fsm-metadata :fsm-seq state-maps)))))
 
 (defmacro defsm-seq